// src/types/Cliente.ts
export interface Cliente {
  id: string;
  nome: string;
  matricula: string;
  dataNascimento: string;
  responsavelFinanceiro: string;
  cpfResponsavel: string;
  valor: string;
  formaPagamento: string;
  telefone: string;
  email: string;
  endereco: string;
  dataCadastro: string;
  dataVencimento: string;
}